update setup.py to make cleaning optional
`setup.py` was configured to automatically remove the compiled SWIG
modules at install time, but this makes it a pain for us when we're
making changes, since every change also requires an install to check,
and every install requires (required...) compiling the SWIG files.

The new behavior is:

> python setup.py install

will install and keep the built SWIG files

> python setup.py install clean

will install and remove the built SWIG files


Former-commit-id: 01c202a3d71199eef369caa91072ba81a8c4d344

+++ setup.py
@@ -25,18 +25,20 @@ class CustomInstall(install):
         self.run_command('build_ext')
         self.do_egg_install()
         #setuptools cleanup is weak, do it manually
-        for tree in ['PyGBe.egg-info', 'build', 'dist']:
-            shutil.rmtree(tree, ignore_errors=True)
-        for swigfile in [
-                'pygbe/tree/calculateMultipoles.py',
-                'pygbe/tree/calculateMultipoles_wrap.cpp',
-                'pygbe/tree/direct.py',
-                'pygbe/tree/direct_wrap.cpp',
-                'pygbe/tree/multipole.py',
-                'pygbe/tree/multipole_wrap.cpp',
-                'pygbe/util/semi_analyticalwrap.py',
-                'pygbe/util/semi_analyticalwrap_wrap.cpp',]:
-            os.remove(swigfile)
+        cmdline = ''.join(sys.argv[1:])
+        if 'clean' in cmdline:
+            for tree in ['PyGBe.egg-info', 'build', 'dist']:
+                shutil.rmtree(tree, ignore_errors=True)
+            for swigfile in [
+                    'pygbe/tree/calculateMultipoles.py',
+                    'pygbe/tree/calculateMultipoles_wrap.cpp',
+                    'pygbe/tree/direct.py',
+                    'pygbe/tree/direct_wrap.cpp',
+                    'pygbe/tree/multipole.py',
+                    'pygbe/tree/multipole_wrap.cpp',
+                    'pygbe/util/semi_analyticalwrap.py',
+                    'pygbe/util/semi_analyticalwrap_wrap.cpp',]:
+                os.remove(swigfile)
 
 def main():
     if sys.version_info[0] != 2:
